Move JsonConfigStorage to ConfigUtil
